requirements.txt
django==4.0.0
djangorestframework
channels==3.0.4
Pillow